+++ src/BabyKusto.Core/Evaluation/BuiltIns/Operators/Divide.cs
@@ -14,7 +14,7 @@ internal class DivideIntOperatorImpl : IScalarFunctionImpl
         Debug.Assert(arguments.Length == 2);
         var left = (int?)arguments[0].Value;
         var right = (int?)arguments[1].Value;
-        return new ScalarResult(ScalarTypes.Int,
+        return new ScalarResult(ScalarTypes.Long,
             left.HasValue && right.HasValue && right.Value != 0 ? left.Value / right.Value : null);
     }
 
@@ -25,7 +25,7 @@ internal class DivideIntOperatorImpl : IScalarFunctionImpl
         var leftCol = (TypedBaseColumn<int?>)(arguments[0].Column);
         var rightCol = (TypedBaseColumn<int?>)(arguments[1].Column);
 
-        var data = new int?[leftCol.RowCount];
+        var data = new long?[leftCol.RowCount];
         for (var i = 0; i < leftCol.RowCount; i++)
         {
             var (left, right) = (leftCol[i], rightCol[i]);
