+++ src/BabyKusto.Core/Evaluation/BuiltIns/ScalarFunctions/Bin.cs
@@ -16,7 +16,7 @@ internal class BinIntFunctionImpl : IScalarFunctionImpl
         var left = (int?)arguments[0].Value;
         var right = (int?)arguments[1].Value;
 
-        return new ScalarResult(ScalarTypes.Int, Floor(left, right));
+        return new ScalarResult(ScalarTypes.Long, Floor(left, right));
     }
 
     public ColumnarResult InvokeColumnar(ColumnarResult[] arguments)
@@ -26,7 +26,7 @@ internal class BinIntFunctionImpl : IScalarFunctionImpl
         var leftCol = (TypedBaseColumn<int?>)(arguments[0].Column);
         var rightCol = (TypedBaseColumn<int?>)(arguments[1].Column);
 
-        var data = new int?[leftCol.RowCount];
+        var data = new long?[leftCol.RowCount];
         for (var i = 0; i < leftCol.RowCount; i++)
         {
             var (left, right) = (leftCol[i], rightCol[i]);
@@ -37,7 +37,7 @@ internal class BinIntFunctionImpl : IScalarFunctionImpl
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static int? Floor(int? left, int? right)
+    private static long? Floor(int? left, int? right)
     {
         if (left.HasValue && right.HasValue)
         {
